+++ src/native/NativeLuciq.ts
@@ -144,6 +144,7 @@ export interface LuciqNativeModule extends NativeModule {
 
   // Report APIs //
   setPreSendingHandler(): void;
+  unsetPreSendingHandler(): void;
   appendTagToReport(tag: string): void;
   appendConsoleLogToReport(consoleLog: string): void;
   setUserAttributeToReport(key: string, value: string): void;