+++ src/specs/NativeLuciq.ts
@@ -117,6 +117,7 @@ export interface Spec extends TurboModule {
 
   // Report APIs //
   setPreSendingHandler(): void;
+  unsetPreSendingHandler(): void;
   appendTagToReport(tag: string): void;
   appendConsoleLogToReport(consoleLog: string): void;
   setUserAttributeToReport(key: string, value: string): void;