+++ src/specs/NativeLuciqBugReporting.ts
@@ -29,8 +29,11 @@ export interface Spec extends TurboModule {
   setShakingThresholdForAndroid(threshold: number): void;
 
   setOnInvokeHandler(): void;
+  unsetOnInvokeHandler(): void;
   setDidSelectPromptOptionHandler(): void;
+  unsetDidSelectPromptOptionHandler(): void;
   setOnSDKDismissedHandler(): void;
+  unsetOnSDKDismissedHandler(): void;
 
   addUserConsent(
     key: string,